Level3 引数化
