--- nn/nn_num/src/auto.sh
@@ -1,7 +1,0 @@
-l
-l
-l
-c
-e
-data.kanji/eva1-2.txt
-q
