+++ nn/nn_num/src/command2.txt
@@ -1,0 +1,2 @@
+l
+q
