+++ nn/nn_num/src/data.kanji/learn1-2.txt
@@ -1,19 +1,19 @@
-3
-00000000000000000
-00000000000000000
-00000000000000000
-11111111111111111
+1
 00000000000000000
 00000000000000000
 00000000000000000
 00000000000000000
 00000000000000000
-11111111111111111
 00000000000000000
 00000000000000000
 00000000000000000
 00000000000000000
-11111111111111111
+01111111111111110
+00000000000000000
+00000000000000000
+00000000000000000
+00000000000000000
+00000000000000000
 00000000000000000
 00000000000000000
 00000000000000000
