+++ nn/nn_num/src/data.kanji/learn1-3.txt
@@ -1,19 +1,19 @@
-3
-00000000000000000
-00000000000000000
-00000000000000000
-00111111111111100
+1
 00000000000000000
 00000000000000000
 00000000000000000
 00000000000000000
 00000000000000000
-00011111111111000
 00000000000000000
 00000000000000000
 00000000000000000
 00000000000000000
 01111111111111110
+00000000000000000
+00000000000000000
+00000000000000000
+00000000000000000
+00000000000000000
 00000000000000000
 00000000000000000
 00000000000000000
