+++ nn/nn_num/src/run_nn.bash
@@ -1,19 +1,19 @@
 #!/bin/bash
 
-# ï¼œæ¦‚è¦ï¼
-# NNã«ã‚ˆã‚‹å­¦ç¿’ã‚’å®Ÿè¡Œã—ï¼Œ
-# ã€€1. å­¦ç¿’ã®æ§˜å­ï¼ˆå­¦ç¿’ã‚¹ãƒ†ãƒƒãƒ—æ•°æ¯ã®èª¤å·®æ¨ç§»å›³ï¼‰ã‚„ï¼Œ
-# ã€€2. å­¦ç¿’å¾Œã®å­¦ç¿’ãƒ‡ãƒ¼ã‚¿ã«å¯¾ã™ã‚‹å­¦ç¿’åº¦åˆã„ï¼Œ
-# ã€€3. å­¦ç¿’ã«ç”¨ã„ãªã‹ã£ãŸãƒ†ã‚¹ãƒˆç”¨ãƒ‡ãƒ¼ã‚¿ã«å¯¾ã™ã‚‹è©•ä¾¡ï¼Œ
-# ã‚’è¡Œã†ï¼ˆäºˆå®šï¼‰ï¼
+# ¡ã³µÍ×¡ä
+# NN¤Ë¤è¤ë³Ø½¬¤ò¼Â¹Ô¤·¡¤
+# ¡¡1. ³Ø½¬¤ÎÍÍ»Ò¡Ê³Ø½¬¥¹¥Æ¥Ã¥×¿ôËè¤Î¸íº¹¿ä°Ü¿Ş¡Ë¤ä¡¤
+# ¡¡2. ³Ø½¬¸å¤Î³Ø½¬¥Ç¡¼¥¿¤ËÂĞ¤¹¤ë³Ø½¬ÅÙ¹ç¤¤¡¤
+# ¡¡3. ³Ø½¬¤ËÍÑ¤¤¤Ê¤«¤Ã¤¿¥Æ¥¹¥ÈÍÑ¥Ç¡¼¥¿¤ËÂĞ¤¹¤ëÉ¾²Á¡¤
+# ¤ò¹Ô¤¦¡ÊÍ½Äê¡Ë¡¥
 
-# ï¼œä½¿ã„æ–¹ï¼
+# ¡ã»È¤¤Êı¡ä
 # prompt> ./run_nn.bash $rand-seed
 
 EXE=nn_num
 
 
-# å­¦ç¿’ã‚¹ãƒ†ãƒƒãƒ—æ•°æ¯ã®èª¤å·®æ¨ç§»å›³ä½œæˆ
+# ³Ø½¬¥¹¥Æ¥Ã¥×¿ôËè¤Î¸íº¹¿ä°Ü¿ŞºîÀ®
 learning_fig() {
     datafile=`echo $STDOUT | sed s/txt/data/`
     figfile=`echo $STDOUT | sed s/txt/svg/`
@@ -27,7 +27,7 @@ learning_fig() {
     gnuplot < .gnuplot
 }
 
-# 2. å­¦ç¿’å¾Œã®å­¦ç¿’ãƒ‡ãƒ¼ã‚¿ã«å¯¾ã™ã‚‹å­¦ç¿’åº¦åˆã„
+# 2. ³Ø½¬¸å¤Î³Ø½¬¥Ç¡¼¥¿¤ËÂĞ¤¹¤ë³Ø½¬ÅÙ¹ç¤¤
 learned_fig() {
     datafile=`echo $STDOUT | sed s/txt/learned/`
     grep CHECK $STDOUT | cut -f2- -d" " > $datafile
