+++ nn/nn_num/src/varParam.h
@@ -1,0 +1,3 @@
+#define ETA 1.9
+#define ALPHA 1.9
+#define HIDDEN 10
